requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
plotly==5.20.0
python-dateutil==2.9.0.post0 # Or a recent stable version, e.g., 2.8.2